cleaned datasets, removed useless comments, renamed some files in models, moved functions to more relevant folders
