+++ competition/tf-scatt&cnn_100sample_cifar10.py
@@ -28,7 +28,7 @@ DF = 25
 THREE_PHASE = 1
 SEED = int(time.time() * np.random.rand(1))
 LEARNABLE = 1
-EPOCHS = 1
+EPOCHS = 3000
 INIT = "Tight-Frame"
 RUNS_PER_SEED = 10
 SCHEDULER = "OneCycleLR"
