+++ parametricSN/models/models_factory.py
@@ -2,15 +2,22 @@
 
 Author: Benjamin Therien
 
-functions: 
+Functions: 
     baseModelFactory -- Factory for the creation of the first part of a hybrid model
     topModelFactory -- Factory for the creation of seconds part of a hybrid model
+
+Exceptions:
+    InvalidArchitectureError -- Error thrown when an invalid architecture name is passed
 """
 
 
 from .sn_base_models import sn_Identity, sn_ScatteringBase
 from .sn_top_models import sn_CNN, sn_MLP, sn_LinearLayer, sn_Resnet50
-from .sn_models_exceptions import InvalidArchitectureError
+
+class InvalidArchitectureError(Exception):
+    """Error thrown when an invalid architecture name is passed"""
+    pass
+
 
 def baseModelFactory(architecture, J, N, M, second_order, initialization, seed, device, 
                      learnable=True, lr_orientation=0.1, lr_scattering=0.1, filter_video=False,
